% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{LungData}
\alias{LungData}
\title{Gene expression profiling of chronic lung disease for the Lung Genomics Research Consortium (GSE47460)
Microarray data from total RNA extracted from whole lung homogenate from subjects undergoing thoracic surgery.}
\format{
a list of four components.
\itemize{
\item{X: }{Expression matrix with 319 samples(Rows) and 2000 genes(Cols) after preprocessing and centering}
\item{Y: }{7 outcomes (Cols) of 319 samples (Rows): "fev1pd1a", "fvcprd1", "bode", "ratiopre", "RV", "WBCDIFF1", "WBCDIFF4"}
\item{Z: }{intercept and covariates of age, gender and BMI for 319 samples}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE47460}
}
\usage{
LungData
}
\description{
Gene expression profiling of chronic lung disease for the Lung Genomics Research Consortium (GSE47460)
Microarray data from total RNA extracted from whole lung homogenate from subjects undergoing thoracic surgery.
}
\keyword{datasets}
