% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{MCMC}
\alias{MCMC}
\title{Metropolis-within-Gibbs Sampler for mgClustBCC Inference}
\usage{
MCMC(
  n_iter,
  K,
  X,
  Y,
  Y.ind = NULL,
  Z,
  alpha_G.init = 0.5,
  alpha_O.init = 0.5,
  v0_vec = 100,
  delta = 0.5,
  d1 = 0.001,
  d2 = 0.001,
  t1 = 2,
  t2 = 2,
  a1 = 4,
  a2 = 450,
  b1 = 2,
  b2 = 0.005,
  r_G1 = 5,
  r_G2 = 1,
  r_O1 = 1,
  r_O2 = 1,
  h = NULL,
  cutoff_G = NULL,
  cutoff_O = NULL,
  init.p.adjust = TRUE,
  init.pcut = 0.05,
  init.NG = NULL
)
}
\arguments{
\item{n_iter}{the number of iteration}

\item{K}{the number of clusters}

\item{X}{N by G omics matrix, where each row is a sample}

\item{Y}{N by O outcome matrix, where each row is a sample}

\item{Y.ind}{censor indicators. NULL if none of the outcomes in Y is survival, otherwise N by O matrix where non-survival outcomes are NA}

\item{Z}{covariates in the outcome association model, N by (q+1) matrix where the first column is the intercept of value 1.}

\item{alpha_G.init}{initial value of omics-specific consensus rate, 0.5 by default}

\item{alpha_O.init}{initial value of outcome-specific consensus rates, a vector of the same length as the number of outcomes, 0.5 by default}

\item{v0_vec}{variance scale of cluster-specific intercepts, a vector of the same length as the number of outcomes, 100 by default}

\item{delta}{step size of Metropolis-Hasting sampler for survival outcome, 0.5 by default}

\item{d1}{shape of inverse gamma prior for omics variance tau, 0.001 by default}

\item{d2}{rate of inverse gamma prior for omics variance tau, 0.001 by default}

\item{t1}{shape of inverse gamma prior for outcome variance sigma (continous or survival), 2 by default}

\item{t2}{rate of inverse gamma prior for outcome variance sigma (continous or survival), 2 by default}

\item{a1}{shape of inverse gamma prior for the variance of cluster-varying genes, 4 by default}

\item{a2}{rate of inverse gamma prior for the variance of cluster-varying genes, 450 by default}

\item{b1}{shape of inverse gamma prior for the variance of stable genes, 2 by default}

\item{b2}{rate of inverse gamma prior for the variance of stable genes, 0.005 by default}

\item{r_G1}{shape1 of truncated beta prior for the omics-specific consensus rate, 5 by default}

\item{r_G2}{shape2 of truncated beta prior for the omics-specific consensus rate, 1 by default}

\item{r_O1}{shape1 of truncated beta prior for the outcome-specific consensus rates, a vector of the same length as the number of outcomes, 1 by default}

\item{r_O2}{shape2 of truncated beta prior for the outcome-specific consensus rates, a vector of the same length as the number of outcomes, 1 by default}

\item{h}{parameters of Dirichlet prior, rep(1,K) by default}

\item{cutoff_G}{truncation point of truncated beta prior for the omics-specific consensus rate, 1/K by default}

\item{cutoff_O}{truncation point of truncated beta prior for the outcome-specific consensus rates, a vector of the same length as the number of outcomes, 1/K by default}

\item{init.p.adjust}{pre-screening based on original p-values or adjusted p-values, TRUE by default}

\item{init.pcut}{pre-screening p-value cutoff, 0.05 by default}

\item{init.NG}{pre-specified number of initially selected genes, NULL by default}
}
\value{
A list of posterior samples.
\itemize{
\item{s.steps: }{n_iter by G binary matrix where each row represents the gene selection values at one iteration}
\item{C.steps: }{n_iter by N matrix of master cluster where each row represents master cluster labels of all samples at one iteration}
\item{L_G.steps: }{n_iter by N matrix of omics-specific cluster where each row represents omics-specific cluster labels of all samples at one iteration}
\item{L_O.steps: }{a list of n_iter elements, where each element is an N by O matrix of omics-specific clusters with each row representing cluster labels of one sample for all outcomes}
\item{BETA0.steps: }{a list of n_iter elements, where each element is an O by K matrix of intercepts across outcome-specific cluster for each outcome}
\item{BETA.steps: }{a list of n_iter elements, where each element is an O by q matrix of covariate coefficients for each outcome}
\item{MU.steps: }{a list of n_iter elements, where each element is a G by K matrix with each row representing the omics-specific cluster centers for each feature}
\item{pars.steps: }{a matrix of parameters, where each row represents omics-specific consensus rate, outcome-specific consensus rates, variance of cluster-varying and stable genes, omics variance, outcome variance, cluster probability, probablity of gene selection for one iteration}
}
}
\description{
Metropolis-within-Gibbs Sampler for mgClustBCC Inference
}
\examples{
\dontrun{
  data('LungData') #load dataset
  test=MCMC(n_iter=1000,K=K,X=LungData$X[,1:1000],Y=LungData$Y[,1:3],
  Y.ind=NULL,Z=LungData$Z)
}
}
