% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{MCMCpredict}
\alias{MCMCpredict}
\title{Predict the cluster labels}
\usage{
MCMCpredict(X = X, Y = Y, Z = Z, estimates, K, Y.ind = NULL)
}
\arguments{
\item{X}{N by G omics matrix, where each row is a sample}

\item{Y}{N by O outcome matrix, where each row is a sample (optional)}

\item{Z}{covariates in the outcome association model, N by q matrix without column of intercept}

\item{estimates}{output of estimate function}

\item{K}{the number of clusters}

\item{Y.ind}{censor indicators. NULL if none of the outcomes in Y is survival, otherwise N by O matrix where non-survival outcomes are NA}
}
\value{
A list of predicted cluster labels.
\itemize{
\item{C.pred: }{a vector of length N, the predicted master cluster label for each sample}
\item{L_G.pred: }{a vector of length N, the predicted omics-specific cluster label for each sample}
\item{L_O.pred: }{a N by O matrix where each row represents the predicted outcome-specific cluster label for each sample}
}
}
\description{
Predict the cluster labels
}
\examples{
\dontrun{
test.pred=MCMCpredict(X=LungData$X[,1:1000],Y=LungData$Y[,1:3],
Y.ind=NULL,Z=LungData$Z[,-1],K=3,estimates=test.est)
}
}
