% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postLogLK.R
\name{postLogLK}
\alias{postLogLK}
\title{Log likelihood by plugging in posterior samples}
\usage{
postLogLK(X, Y, Z, MCMCres, nsamples = 1000, Y.ind = NULL)
}
\arguments{
\item{X}{N by G omics matrix, where each row is a sample}

\item{Y}{N by O outcome matrix, where each row is a sample}

\item{Z}{covariates in the outcome association model, N by q matrix without column of intercept}

\item{MCMCres}{output of MCMC function}

\item{nsamples}{the number of posterior samples after burn-in (must smaller than n_iter in MCMC function), 1000 by default}

\item{Y.ind}{censor indicators. NULL if none of the outcomes in Y is survival, otherwise N by O matrix where non-survival outcomes are NA}
}
\value{
the median of log-likelihood over all posterior samples
}
\description{
Log likelihood by plugging in posterior samples
}
\examples{
\dontrun{
test.loglkh=postLogLK(X=LungData$X[,1:1000],Y=LungData$Y[,1:3],
Y.ind=NULL,Z=LungData$Z[,-1],nsamples=500,MCMCres=test)
}
}
