% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{TNBC_Metabric}
\alias{TNBC_Metabric}
\title{Gene expression profiling of METARBIC breast cancer dataset, 271 triple negative breast cancer samples are picked.}
\format{
a list of five components.
\itemize{
\item{X: }{Expression matrix with 271 samples(Cols) and 2000  genes(Rows) after preprocessing and centering}
\item{Y: }{3 outcomes (Cols) of 271 samples (Rows): "Tumor_size", "lymph_nodes", "OS"(survival) }
\item{Y.ind: }{censoring indicators for survival outcome where 1 means events and 0 means censoring}
\item{Z: }{intercept and covariates of age and Chemotherapy for 271 samples}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE47460}
}
\usage{
TNBC_Metabric
}
\description{
Gene expression profiling of METARBIC breast cancer dataset, 271 triple negative breast cancer samples are picked.
}
\keyword{datasets}
