% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{TNBC_ScanB}
\alias{TNBC_ScanB}
\title{Gene expression profiling of SCAN-B breast cancer dataset, 141 triple negative breast cancer samples are picked.}
\format{
a list of five components.
\itemize{
\item{X: }{Expression matrix with 141 samples(Cols) and 2000  genes(Rows) after preprocessing and centering}
\item{Y: }{3 outcomes (Cols) of 141 samples (Rows): "Tumor_size", "lymph_nodes", "OS"(survival) }
\item{Y.ind: }{censoring indicators for survival outcome where 1 means events and 0 means censoring}
\item{Z: }{intercept and covariates of age and Chemotherapy for 141 samples}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE60789}
}
\usage{
TNBC_ScanB
}
\description{
Gene expression profiling of SCAN-B breast cancer dataset, 141 triple negative breast cancer samples are picked.
}
\keyword{datasets}
